enid/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==4.2.7
django-autoslug==1.9.9
django-cors-headers==4.3.1
django-seed==0.3.1
djangorestframework==3.14.0
Faker==20.1.0
gunicorn==21.2.0
idna==3.6
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==7.12.0
toposort==1.10
typing_extensions==4.8.0
urllib3==2.1.0
